Update config docs-in-comment.
I'm running Grocy in Docker behind a reverse proxy, and when looking to adjust the URL so I wasn't redirected to `localhost`, I skipped over this setting at first because of the edited language.

Please consider this a starting point. I didn't want to overstep, but I think this could be adjusted further to explain that this value should be the protocol and domain Grocy is hosted at.