Replace generic object api in openapi schema with specific ones
Change OpenApiController so it copies /objects/{entity} path for each
entity type ant removes {entity} parameter replacing it with a antity
name being path of the path.

Instead of a anyOf return type containing all possible entities each
endpoint has a type specific for this entity. Note: only for entities
that already had a schema defined for the return value.