Make date in backup filename ISO 8601 compliant
By making the date in backup filename ISO 8601 compliant (year-month-day) an alphabetical sort of the backups automatically will be a chronological sort of the backups. This is more convenient.

As the automatic deletion of backups older than 60 days relies on the modified date of the file (and not the filename), this will continue to work without problems.

See also: https://www.xkcd.com/1179/