Fix Location Content Sheet filter checkbox
Previously, the query parameter functioned as expected, but the
checkbox always appeared checked. Additionally, clicking the checkbox
would always result in the "unchecked" state despite remaining visually
checked.

This adds the required context to the controller and an @if directive to
the checkbox's checked attribute in the view to match.