Get the image file type from the response content type (#2811)
* Get the image file type from the response content type
When using the external lookup tool, the file extention in the URL is used for the name of the downloaded image. However, some API's do not use file types in the resource name. If that is the case, the Content-Type of the request will now be used to define the file extention.

* Typo: Deleted last empty line on accident.

* Typo

* Apply code style and simplify this

---------

Co-authored-by: Bernd Bestel <bernd@berrnd.de>